requirements.txt
flask
python-dotenv
slack_sdk
slackeventsapi